pyproject.toml
[project]
name = "contactlistappseleniumautomation"
version = "0.1.0"
description = "ContactList_UI_API_test"
authors = [
    {name = "Sandro Jimena sjimena@che2systems.com"}
]
license = {text = "TE"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "allure-pytest (>=2.13.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "serial",
]
